Add held_span, don't be so vectory
This should cut down on the amount of copying needed to move things between
cores.

Also, deleters don't know their cores by default, so you have to use foreign_ptr
in building them. But it's still worthwhile since you don't have the detail of
how to release it in the type.